fix: active menu font style 적용 오류 수정
